+++ ChangeLog.md
@@ -14,6 +14,8 @@
 
 * Fixed bug in default `Sampler` implementations which weren't generating initial starting points with maximal randomness.
 
+* Fixed bug in default pytorch `Sampler` implementation where PGD search wasn't being run correctly if the loss function was being called in a `torch.no_grad()` environment.
+
 * Fixed bug where constraints with multiple quantified variables in them were being incorrectly added to the sampler domains.
 
 * Quantified variables are no longer required to have a well-defined domain.